replace graphiql with playground, add docker-compose with postgresql and init scripts

+++ modules/core/src/main/scala/model/City.scala
@@ -9,5 +9,5 @@ final case class City(
   name:         String,
   countryId:    String,
   district:     String,
-  population:   Int,
+  population:   Long,
 )
